+++ cornerstone-system/src/main/java/com/cornerstone/system/service/impl/SysOperLogServiceImpl.java
@@ -2,7 +2,7 @@ package com.cornerstone.system.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
-import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.baomidou.mybatisplus.spring.service.impl.ServiceImpl;
 import com.cornerstone.system.domain.entity.SysOperLog;
 import com.cornerstone.system.domain.mapper.SysOperLogMapper;
 import com.cornerstone.system.service.SysOperLogService;